refactor(core): Extract update_workflow MCP tool schemas into their own module (no-changelog)
Moves the zod input/output schemas, the tool description builders and the
loose-to-strict operation parse out of update-workflow.tool.ts and into a
new partial-update-schemas.ts sibling (1073 -> 773 lines).

Pure move: the relocated code is byte-identical apart from the comment
about buildInputSchema's concrete return type, which sat above the wrong
declaration and is now attached to the function it describes.
